+++ container-engine-lib/docs/changelog.md
@@ -3,7 +3,7 @@
 * Migrated the `engine functions` from `docker_kurtosis_backend` package to the `engine_functions` package
 * Removed duplicated shared functions that where present in the both the `docker_kurtosis_backend` package and the `shared_helpers` package
 * Added the `consts` package to group the general constants used in the Docker Kurtosis backend
-* Added the `ForLogsCollectorServer` Docker objects attribute
+* Added the `ForLogsDatabaseServer` and the `ForLogsCollectorServer` Docker objects attribute
 
 #### Breaking Changes
 * Move all `docker.DockerKurtosisBackend` functionality to `docker_kurtosis_backend` module